requirements.txt
# Dev environment
autopep8==2.3.2

# App
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
Flask-SocketIO==5.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pydantic==2.11.2
pydantic_core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
rsa==4.9
simple-websocket==1.1.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
websocket-client==1.7.0
pymongo==4.7.2

# Extra Tools
Flask-Dance==7.1.0
oauthlib==3.2.2
requests-oauthlib==2.0.0
urlobject==2.4.3
pymupdf==1.25.5
Pillow==11.2.1
defusedxml==0.7.1
fonttools==4.57.0
fpdf2==2.8.2
narwhals==1.35.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
plotly==6.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
